Eu te amo Mummi!! Feliz 62 anos!!
at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝

   ----
                        -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ os/win_window.cpp
@@ -4531,8 +4531,8 @@ ExitModal:
             {
                if (nCmdShow != SW_HIDE)
                {
-                  ::SetWindowPos(get_safe_handle(), 0, m_rectParentClient.left, m_rectParentClient.top,
-                     m_rectParentClient.width(), m_rectParentClient.height(), SWP_SHOWWINDOW | SWP_NOZORDER);
+                  ::SetWindowPos(get_safe_handle(), 0, (int) m_rectParentClient.left, (int) m_rectParentClient.top,
+                     (int) m_rectParentClient.width(), (int) m_rectParentClient.height(), SWP_SHOWWINDOW | SWP_NOZORDER);
                }
             }
             ::ShowWindow(get_handle(), nCmdShow);
